nassh: make all links clickable via OSC-8 link sequences
Most links we've been showing require the user to copy & paste or to
know to Ctrl+Click them to open.  Add a helper to generate the OSC-8
sequence, then wrap all links we show to the user in that.  Also cut
over existing ad-hoc OSC-8 sequences to the new helper.

Change-Id: Iff550411282a78991367e9bc2ef6e2866e6f96dc
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2116829
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>