terminal: support user variations to themes and reset
New pref theme-variations is an object which
stores the set of prefs which have been modified
for each theme.

When a theme is modified from its defaults, we
record which values are modified so that they can
be reapplied if that theme is selected again. We
also display the reset icon so that a theme can
be reset to its defaults.

Refactored dialog styles from colorpicker to be reused
for the reset confirmation.

Bug: 1048468
Change-Id: I305c0ee06230db81c43de87e319ef6919e766406
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2120655
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jason Lin <lxj@google.com>