terminal: store colorpicker values as uppercase hex
When colorpicker hex input is used, store the resulting
color as uppercase hex rather than rgb since defaults
are all registered as uppercase hex and this allows
us to correctly detect when a value has been reset
to its default value.

This is important for terminal themes reset which
needs to correctly detect when values are using
defaults.

Also updated normalizePrefs to store as upper hex,
and display to always be upper.

Bug: 1048468
Change-Id: I842955b54b5ac4b1cda53b779ac7e271566c5737
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2120654
Reviewed-by: Jason Lin <lxj@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>