nassh: api: switch to using a Map
This avoids dealing with JS objects as maps and the insecurity of all
that by using a proper Map object.  Not that this code was known to
be buggy, just being defensive here and using good coding patterns.

This requires a little bit more boilerplate, but I think it's worth
it for the clarity & stability in lookups.

Change-Id: Ibbf20881bb3a55b9274c3846ed84b074728e1ce3
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2125628
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>