nassh: add a helper page for registering protocol handlers
The current flow for registering protocol handlers is not great.
We often are loaded in such a way that we don't actually get to
prompt the user (because Chrome doesn't allow it as it doesn't
detect that the call came from a user intent).  Lets add a page
so users can read & understand what's going on, and then click
some buttons so Chrome knows it's OK to register things.

Bug: b/152554978
Change-Id: Ie2aa4ac992666bfca190242b97f02d8116012c2c
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2125629
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>