nassh: improve translation import process
When we modify strings on our side, it forces all translations to go
stale.  We'd rather have slightly stale messages than completely use
untranslated English messages.  So check the new messages to see if
any match the English baseline.  If so, restore the old message.

If there is no old message, it means it hasn't been translated yet,
so we still want to discard it.  Chrome will fall back to English at
runtime itself, so there's no need to waste space on it.

Change-Id: Id7e02f7c2019602c76bd5fa1eb85935b07b76642
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2125482
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>