nassh: workaround bug in translation pipeline
Our translation pipeline doesn't like it if you use the same word in
all uppercase as the placeholder field.  i.e. It dies because it sees
"UID" and "$UID$" and thinks the first (plain string) is part of the
placeholder when it's not.  Rename the placeholders in this message
to workaround the bug and unblock translations.

Bug: b/152714834
Change-Id: I6bc80a205b5d0d98698a7b849dbd1e743d96c399
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2128729
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>