nassh: streams: throttle WebSocket sending as needed
The web platform says that trying to send data via a WebSocket whose
buffers are full (and unable to queue the data) will result in the
socket being closed with an error.  This not ideal.

Unfortunately the platform doesn't seem to have any APIs to find out
what the limit is, or to have an error be returned directly (so that
we could mitigate things directly), so the best we can do is add some
checks ahead of time and hope for the best.

There isn't a way to query the current platform's limits, so we pick
64KiB somewhat arbitrarily.  Chrome seems to use 128KiB as its own
limit at least, and it's twice as large as the largest relay frame we
support currently, and normally the platform is able to queue/drain
things fast enough already.

We haven't had reports of this, but it seems like something we should
account for just in case, especially as a recent bug in relay handling
implicated this logic.

Change-Id: I3a1a110839f853b17fb43b2c952eb1124c2be2e1
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2146146
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>