nassh: corpv4: fix plugin ack logic
There are two acks in play here:
(1) When NaCl writes to JS, JS needs to ack back to unblock NaCl so
    it can unblock its write() syscall.
(2) When JS writes data to the relay, the relay will ack back so the
    client knows it can discard bytes from its queue.

Currently the corpv4 stream doesn't send the NaCl ack until the relay
has sent us back an ack.  The other stream classes will send the NaCl
ack as soon as they've written the data to the socket.  This ends up
reducing bandwidth speeds significantly due to the latency between the
clients.  Adjusting the corpv4 stream to match the other streams and
increases speed from ~150KB/s to ~2.7MB/s on my home network.

Change-Id: I8a08f589baf3e14399e810f550d30a8e326c25fd
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2146148
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>