nassh: buffer: new API for managing byte buffers
We have a number of scenarios where we accept fragments of data (as
ArrayBuffers) that we need to merge into a single (logical) buffer
before we can use it.  Define a new buffer API to transparently
support that use case.

The current relay streams utilize lib.array.concatTyped for writes
from the NaCl world.  This helper will create a new ArrayBuffer that
is big enough for both inputs, copy them in, and then the two inputs
are marked for GC.  When the NaCl side writes a lot of data, this
overhead of creating a lot of ArrayBuffers & memcpying can start to
add up in CPU utilization, and given a fast enough link, might start
to slow down the transfers.

This new buffer API uses that lib.array.concatTyped logic, so it
should be mostly code shuffling at this point with no real change
in overall behavior.  We'll add alternative buffer implementations
for live testing/comparison as time goes on.

This can be controlled via --field-trial-buffer=<implementation>.

Change-Id: I7786dc52329a6328b9344403ecdaca57d5331942
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2148832
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>