nassh: buffer: add a scatter/gather buffer
Add a scatter/gather class that avoids creating new temporary buffers
and the resulting memcpys by keeping references to the original inputs
and maintain pointers/counters to walk them.  So when NaCl writes to
us, we queue its ArrayBuffer, and when we send data to the relay, we
manually walk those buffers to keep track of how much has been read.
When the relay acks back, we walk the buffer list to see which ones we
can now discard.

Change-Id: I0f182f8aa5ae6c13bfda8976733e93fd5670b68d
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2148833
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>