nassh: add a basic framework for field trials
With the new buffer framework, we want to be able to toggle features
on the fly for comparison & performance testing.  Start a basic field
trial framework to control these things.  For now we keep it as a set
of command line flags as that's the simplest, but in the future we
could switch these to a UI like chrome://flags.  TBD.

Change-Id: I21c913353826700893391d2fd2f9ef1d7ef17e25
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2146071
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>