nassh: fix identity refresh when importing
The promises logic with files forgot to call resolve() with the new
promise created on the fly.  It was trying to return the promise from
the file operation directly.  Rewrite the code to just queue the file
promise directly.

Also make sure to clear the files list once we're done with it.  If
the user deletes files and then tries to re-import, but the set of
files were the same as before, the change event wouldn't fire.  Not a
super common scenario, but I hit it while testing.

Bug: chromium:1040901
Change-Id: I81d2371be92f30e2f83f9de81c23bca2abc9389a
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2151863
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>