nassh: prefs: fix unselected panels bleeding through
If the selected settings page does not fill up the entire height,
then unselected settings pages below may be interacted with.  You
don't notice it at first as the pages are all marked with opacity
of 0%, but Chrome still allows you to interact with elements you
can't actually see.  Forcing the selected panel to fill the entire
page avoids that by covering up the lower content.

Change-Id: Id4fda4630c6f3c4524433bbea5e20cda2c972385
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2153388
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>