hterm: keep room for scrollbar
At least on CrOS, the system tells us that scrollbar
is zero width. As a result, we size rowNodes to be
so wide that it clips part of the scrollbar, and the
scrollbar is very hard to target and drag.

This ensures that we always leave at least 12px on the
right margin when the user has scrollbar visible.

Bug: 968618
Change-Id: Ia7761f77f62c1dee672a164cdfa18a7e61c72901
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2157411
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>