nassh: remember & restore window dimensions per-profile
The extension has always created windows at 0x0 with 900x600 size,
but users often move & resize it which is tedious.  Chrome would
handle remembering things on a per-app basis, but only if we don't
specify any specific window settings which doesn't have a good
initial experience, and that isn't per-profile.

Have nassh track the current window dimensions as they change (or
at least as best we can), and restore them when starting up.  We
only do this for extensions as the experience with the native app
can be a bit jarring when switching between the connection dialog
and the saved profile.

Bug: chromium:926561
Change-Id: I3d73bc939d31201183970d9e98b04088cd186232
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2166519
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>