hterm: hide context menu for most key presses
We automatically hide the context menu when clicking the mouse outside
of the menu, but not when the user starts typing.  Add some logic to
handle that case, but be sure to not hide it when only pressing keys
like Alt/Ctrl/Meta by themselves since Ctrl+Click is how you open it!

Bug: chromium:899293
Change-Id: I29cc3ab6158c898f3afc41143011ae8c8cfe4cae
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2166520
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>