nassh/terminal: add Send Feedback buttons everywhere
We link to documents explaining how to file feedback/bug reports,
but lets add dedicated "Send Feedback" buttons everywhere to help
users discover this.  For now it just takes them to the public
issue tracker, but this lays the groundwork for producing real
feedback reports in the future if we want.

Change-Id: Idbf5745e529b6b7f0d2519419ad6dcdb9766d49e
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2174975
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>