libdot: fetch: add a 5 min timeout
The default timeout is no timeout at all.  Change it to 5 min because
if the initial open/handshake or reading any data at all takes longer
than that, something is seriously wrong.

This is *not* the timeout for the entire transfer, just for each step
along the way.  So it's not like we require people to have a connection
that is able to download 200 MiB in <5 min.

Change-Id: Ia5ed731fecedb4538298403931b6efded31b6797
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2176793
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>