hterm: handle padding for fold and a11y buttons
Selected nodes are moved above and below the fold when scrolling with
a selection.  These now become visible when there is padding in the
screen.  The solution is to make the folds equal height to the padding.

A11y buttons also need to adjust their placement to compensate for
padding.

Bug: 1081462
Change-Id: Iafc9f288e6adedd331d688a8047d7582eb3113ad
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2195221
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>