terminal: remap Ctrl+N to call terminalPrivate.openWindow
It is not possible to use web APIs to open a new tabbed app window.

window.open will open a new tab, or a popup.
chrome.tabs.create will open a new tab
chrome.windows.create will open a normal or popup window, but not app.

The web platform team are talking about adding or extending web APIs to
support this, but until then, we must use a private API.

Bug: 1082006
Change-Id: I49719df29dfcc0062b96aba0a5f8dacf0d9a339a
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2200497
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>