nassh: refresh stable icons
Refresh the stable icons by changing ">BETA_" to ">SSH_" as we are
probably out of beta status for a while now.

Create some SVG's of the new images to make it easy to scale to any
size we need.  The ones in here previously were created by hand so
there weren't any masters before.  We provide a few for scale:
* 16.svg: Hand tailored for the 16px icon to avoid anti-alias noise.
* 48.svg: Simpler icon 24px - 64px range.
* 512.svg: The full icon for 64px+ icons.

For the smaller images, use red/yellow rather than blue/green as they
have better contrast ratios on the black background.

Lets ship more resolutions that CWS recommends: 48px for Chrome's
extensions page, and 24px for the extension icon.

Change-Id: Idc7c5685f8f408b10c81913b9d0582e1932fc711
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2214376
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>