nassh: automatically check & refresh SSH certificates
Check with the GCSE extension to see if the SSH certificate is recent.
If it isn't (less than 1 hour before it expires), show a message to
the user that we're kicking off a refresh, and then send a message to
GCSE to do the actual refresh for us.

This logic is currently specific to Google/GCSE, but we could rework
it at some point to allow the user to plugin their own extension.

Bug: b/63900358
Change-Id: Ib29973087a2f8681b25e4908524f69dcf21e46d7
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2219426
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>