nassh: connect: make focused inputs less confusing
We were relying on the configured terminal cursor color to indicate
which input was selected/active as a way of having the connection
dialog integrate nicer with the user's theme.  Unfortunately, since
the default cursor color is red, and users tend to associate that
with "the active field is invalid", they get confused.

Use the default foreground color instead.  If the user set that to
red, then a lot of the UI here will be red in general, so hopefully
they won't associate it directly with "red is invalid".  We still
have the unfocused invalid fields in a red border.

Change-Id: I173626a7df007fa517b1e33c0b81b03c9bb7e91f
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2241139
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>