libdot/hterm/nassh/terminal: rewrite lib.init from callbacks to promises
This pulls the lib.init API up to be promise/async based rather than
the older callback style.  In practice it doesn't matter that much,
but it allows callers to write a bit cleaner code if they don't want
to create a giant init callback func.

We convert a few over to the async methods to prove it works, but
others we do a simpler conversion to .then() out of laziness.

Change-Id: I5ce945919765e899d61ef0a636803b9d93643d67
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2244717
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>