libdot: add helpers for creating archiving
This pulls the existing logic out of node_modules_create_bundle for
creating archives and into the common libdot.py, and improves the
creation & unpacking logic to fix permissions & ownership.

We also clean up the code a bit to use pathlib for paths, and apply
that to the existing unpack helper to keep them consistent.

Change-Id: I43e9ca25a77fc38f88db497f0cef4274863597bf
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2255623
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>