libdot: storage: memory: simplify JSON serialization logic to fix change events
The Local storage backend is only able to handle strings.  Anything
added to it that isn't a string is run through toString() first.
This makes it quite hard to store numbers/arrays/etc...  That is
why the code runs all inputs through JSON.stringify/JSON.parse.

The Memory storage backend, being pure JS, has no such fundamental
limitation.  But we force it to go through JSON anyways so we can
use it in tests to closely mimic the real Local storage behavior.

However, we only really need to do this to normalize the value.
There is no need to internally store it as a JSON string.  So we
can move that normalization logic to the set function only.  This
allows us to simplify the other code paths by keeping the JSON
logic in one place ...

... and it fixes bugs with synthesized change events.  While the
Local storage backend will deserialize when constructing things,
that logic was missing from the Memory backend, which meant we're
sending change events to observers as the raw JSON string.

Change-Id: I5cc4d039831828bc083f7e9b9bac769a4b6a7a7f
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2269585
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>