libdot: storage: memory/terminal: fix missing change events for clear/remove
The Local & Chrome backends generate events when items are removed
from storage, but our Memory & Terminal backends were not.  Add an
internal helper function to both to make it easier to do so.

The Terminal backend was trying to generate change events via its
global onSettingsChanged_, but in practice that never worked: that
callback would fire only after we already updated the cache, so the
new state would match the cache (by design), so we'd never generate
any events.  The code was working for when state changed outside of
our context in the system, but that rarely happens.

Change-Id: Idf1e0b4f549333762c2e63adefba57b88e31542f
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2269587
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>