hterm: update find bar results
The find bar results are updated when text is modified in terminal.
If selected row is modified the first result of the row is selected.
If no results are present on the row we invalidate the selected result
and total number of matches are shown instead.
E.g. :- '3 matches' instead of showing the selected ordinal
E.g. :- '1/3'.
The position of selected result depends arrow buttons pressed on
FindBar.
When up arrow button is pressed the match just before the cursor is
selected.
Similarly, When down arrow button is pressed the match just after
the cursor is selected.
In both cases, the counter label is reverted to it original state and
shows the selected ordinal E.g. :- '1/3'.

Change-Id: I9c6930d305a85b9f23396076030889aa94f57911
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2305810
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>