libdot: jsonlint: new tool for checking style
Make sure the style in our JSON files stay consistent.  We largely
follow JS style guide in that it uses two space indents, so enforce
that on most.

For messages.json which is a bit more data oriented, we use tabs
to save on space and keep it sorted to ease diffing.

We fix the bad style in messages.json, and tweak the style in the
manifest to fit the tool.

Change-Id: I53b8a4a08a1d48e41ec319d28732325554cdeebd
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2320471
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>