nassh: prefs: fix color input syncing
Make sure we only set the color element to #rrggbb as the platform
specifically does not support alpha and will warn+reset itself if
it's specified.  When we changed the lib.colors.rgbToHex API to
return the alpha if the input contained the alpha, this broke as
we were relying on rgbToHex always stripping the alpha away.

Bug: chromium:1113487
Change-Id: Ib8075b705a8bf9dde931919d8a7792e2fa5fefdd
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2405933
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>