nassh: Work around Yubikeys appending arbitrary byte to ECC OIDs
Retry recognizing an ECC curve OID returned by the OpenPGP applet of a
hardware token with the last byte stripped of if the token is a Yubikey.

Yubico has confirmed that Yubikey 5s with firmware version below 5.2.8
append an additional byte to all ECC OIDs. This byte has been assumed to
always be 0 in d098417ebb929251bcf7e48665004dd8a76587ac, when in fact it
can take arbitrary values.

Bug: 1120933
Change-Id: I00c7297631b37911ed84b2e9ff9289a1cebd2738
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2409293
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>