terminal: make sliders keyboard operable
We have two slider elements <transparency-slider> and <hue-sldier>.
Instead of adding more duplicate code (and test code) for each of them
to support keyboard operations, a new slider <terminal-slider> is added.
<transparency-slider> and <hue-slider> will use it and customize the
appearance.

This CL also change the event name "updated" to "change" because the
latter is what native <input> elements use.

This CL also fixes the bug that <transparency-slider> in color picker is
not using the current color but some color with a particular hue. The
knob also follows the current transparency now.

Bug: 1119678
Change-Id: I2d18406be2ede46e59d47ce9d6eb898559d1b29a
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2458941
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>