libdot: colors: delete colorPalette global variable
This is only used by hterm to maintain transient user state that is
actually per-terminal.  Stop using this as a single global variable
by adding a terminal-specific variable to hold overrides instead.

Some colorPallete uses can be switched to the stockPallete as we only
need general info like how big the palette is, not its contents.

Change-Id: I5e9ac3ce66fb04a319fd363396111f9f964b641e
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2491428
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>