hterm: inline global hterm.defaultStorage
This is only used to initialize hterm.PreferenceManager, and it'll
be easier to convert to ES6 modules when we kill off global state.
Make it into an option for hterm.Terminal so people can override
if they want, otherwise we always default to local storage.

We leave the fallback to hterm.defaultStorage until we can finish
migrating nassh & terminal over to passing things down themselves.

Testing requires a bit more effort to specify memory backed storage,
but we don't have much choice here as we wouldn't be able to easily
monkey patch it in a centralized way when using ES6 modules.

Change-Id: I5ff1e696a07b46c98329cbc1890aeb56d8dd439c
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2494363
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>