nassh: fsp: move mounting to API layer
This works by getting a handle to the background page, then invoking
objects on that handle, and poking state directly.  This is a little
funky in that it looks like the code is running in the foreground,
but some resources are actually being associated & assigned to the
background page.  This has been the source of bugs & confusion in
the past.

More importantly, this makes our lives more difficult as we try to
use ES6 modules, get ride of the 'var nassh' decl, and automate SFTP
remounting.  So lets rework the logic so the code execution all
happens in the background thread, and we only plumb back UI messages
to the foreground thread.

Change-Id: I64029110f146f704f3d3b1aaf0561d36da091bbb
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2525781
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>