nassh: connect: improve UI for incomplete connections
We highlight all incomplete fields in red, but users seem to keep
missing that.  They see the disabled "connect" button and think the
program is broken.  We have many reports of this.

Change the UI so the buttons are always enabled, but if you click on
them when the form is incomplete, use the browser's built-in UI to
highlight the exact field that still needs to be completed.

Change-Id: I501e7172cf1ed76cdd8295353a4aa841039a7a11
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2525030
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>