nassh: api: tweak default style for new window helper
The closure compiler cannot handle our dynamic types and local helper
function which processes the defaults.  It's not exactly obvious with
the settings we use currently, but it throws errors when we try moving
the types out into externs.

We can largely avoid this by using the nullish coalescing operator.
This also simplifies the code quite a bit.

Change-Id: I8021f8f256b4daf9b33e27627ce9e52de8c4ea93
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3252003
Reviewed-by: Emil Mikulic <easy@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>