Reland "terminal: open vmshell with the cwd of the parent terminal."
This is a reland of 96ddaa66a12f528d386849819c760b89d362446e with fixes.
The original CL causes bug 1150941 and 1151074. The reason is that the
original CL changed the way we listened to `onbeforeunload` event, but
it did not changed the way we removed the listener, which caused the
listener to not be removed correctly, which ultimately caused the
above-mentioned bugs. In this CL, we revert the `onbeforeunload` code
back to the old version.

Original change's description:
> terminal: open vmshell with the cwd of the parent terminal.
>
> The parent terminal is the last active terminal of the same window just
> before the current terminal is opened. We introduce a new class
> `ActiveTerminalTracker` to track active terminals.
>
> Change-Id: I22b5084589d37418f3c296f60f8cced479cff624
> Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2533067
> Tested-by: kokoro <noreply+kokoro@google.com>
> Reviewed-by: Joel Hockey <joelhockey@chromium.org>

Bug: 1113207
Test: manual tested by entering "exit" in terminal, and tast `crostini.IconAndUsername.buster_amd64_stable`
Change-Id: I92cb29f15112948c6e3295db8ad22a3828e96682
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2550095
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>