ssh_client: add missing exit broadcast
The exit implementation sent a message to the JS world and then went
to sleep waiting for a response, but the response never woke it up.
If another message came in that happened to kick the mutex, we'd see
it, but otherwise we'd wait forever.

Not a big deal at this point, but easy to fix too.

Bug: https://crbug.com/990181
Change-Id: I4f46115a7ad19ccdf40a8dd3b4cc468a9c2c00d4
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2563097
Reviewed-by: Collin Moore <moorect@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>