Call terminal private api only after terminal id is ready
This CL makes sure that we only call terminal private api with a proper
terminal id. Before this change, we might call the api (e.g.
terminalPrivate.onTerminalResize) with the "start up" id, which is not
recognized by the api and causes an error. Also, as a security practice,
we should kill the renderer if an invalid id is received, so we should
make sure the JS code does not use the "start up" id.

Bug: 2519182
Change-Id: I0451b283357d54259070d713ae2f72d8f17ad7fa
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2513956
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>