nassh: fsp: fix mounting after UI rewrite
Two changes landed that independently were fine, but together broke.
We moved mounting to the background page, and we changed the user
input mechanism to use real UI rather than stuffing the terminal.
Since the background page doesn't have access to the foreground
document, it can't generate & render the UI needed.

Rework the existing readpass API a bit so we can wrap & reuse it,
then add a new message as part of the mount API.  The background
page now sends a message to get user input, and the foreground page
takes care of actually rendering & processing that request.

Bug: https://crbug.com/1157584
Change-Id: Ia9279b771ead8b84fd214886c148509c79d261ac
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2586149
Reviewed-by: Collin Moore <moorect@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>