hterm: auto scroll for selection if mouse moves off rows
While primary mouse button (usually left) is down, if the mouse moves
off rows (into padding, margin, or off the window), then auto scroll the
screen in the direction of the selection.

Scroll happens every 200ms and starts by scrolling 1 row, and
accelerates by 20% for each invocation. Starting the scroll slow helps
to make it easy for users to be accurate if they only desire to scroll a
small number of rows, and accelerating helps when users want to scroll
many rows all the way to the end.

Bug: https://crbug.com/865052
Change-Id: I63a9ed42aa917a932fa47753c35269c37bed1291
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2603074
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>