crosh: Use terminalPrivate.openOptionsPage() if running as SWA
When crosh is running as a System Web App, we do not want it to use
regular tabs in ash-crhome.  It will be launched in a tabbed SWA frame
just like Terminal, and it should use the Terminal settings page rather
than calling window.open() and showing the nassh settings page.

Bug: 1161413
Change-Id: I2bc2664ee0cba213f34b1ba8157a4c3702d7bbd6
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2706318
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>