terminal: run crosh from terminal.html
Now that crosh is using terminal settings page, we must load crosh with
the same set of defaults for all prefs.  The simplest way to do this is
to run crosh out of terminal.html.

When terminal.html and the associated JS detects that it is running as
chrome-untrusted://crosh, it will call
terminalPrivate.openTerminalProcess() which opens crosh rather than
terminalPrivate.openVmshellProcess() which opens vmshell.

Change-Id: Iccbb06e81546c1136daee35b05616ee45141da41
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2747895
Reviewed-by: Jason Lin <lxj@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>