terminal: make sure dialogs don't have scrollbars
In the color picker dialog, when you drag a slider's knob to the far
right, a horizontal scrollbar appears on the dialog. This is because we
position the knob beyond the width of the slider. The easiest way to fix
this is just to prevent all dialogs from showing scrollbars at all.

Fixed: 1187975
Change-Id: I386535ef746e843e7792dd8e0ad0841c0bd3e21f
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2758539
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>