nassh: move reconnect dialog to overlay popup
Currently the reconnect dialog is shown in the middle of the terminal
wherever the cursor happens to be.  Depending on the terminal state
when the failure happens, the message might be garbled or lost, or
really hard to pick out from surrounding text.  Move it up to the
overlay that is a distinct UI surface from the underlying content so
it's easily visible & clear when the failure happens.

Change-Id: I48b8b595bd641d1d53fbc275ebec3892e8364a7c
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/2991092
Reviewed-by: Collin Moore <moorect@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>