hterm: adjust text measurement again
The old version used X as the reference glyph for calculating the
width & height.  When the ruler was rewritten to use a canvas, the
glyph was changed to the U+2588 Full Block character assuming it
would always completely fill the space.  Unfortunately, if the
active fonts don't have a monospace box, the calculation could be
wildly off.  Switch back to using X for the width, and include it
with U+2588 to get the max height between the two.

Change-Id: I84f2ff1aa9d61850f9a146884b1bf5ce7ed8ab95
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3193121
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>