nassh: nasftp: rework command line completion for deeper integration
Rework the way we completion command lines to support more than just
the command name (the first argument).  Support completing options,
and allow commands to control completion of their sub-arguments.  We
have to rework from a synchronous API to asynchronous in order to
make this truly useful (querying the server state), so switch the
APIs over to async.

For now, this is mostly groundwork.  Completion of the help command
is included to show the framework working.

Change-Id: I142d13fe3d0639ceb7378c2cd45487c14ffae723
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3207649
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>