hterm: create hterm.Terminal.DEFAULT_PROFILE_ID constant
Refactor only change to create a constant for 'default'.

I was looking to add a hterm.PreferenceManager('default') in
nassh_connect_dialog.js so that it can immediately update any theme
changes. I probably wont go ahead with that change, but it still makes
sense to consolidate usages of 'default'.

Change-Id: Ib4a8975eff2cfc4f7e7e27ae03541fbdad52d595
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3206978
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>