nassh: move nassh.loadWebFonts out of critical path
The popup was running this code before it initialized which seems to
be causing load issues for users with slow/intermittent networks.
Since webfonts aren't important to this UI, just drop it entirely.

For the connection dialog, delay its loading until after we've created
the dialog so we know things are usable.  This largely matches how the
fonts are loaded in the actual terminal now.

Bug: https://crbug.com/1160931
Change-Id: I358bb9ea6433883f0a181af25ffd7ff78f22777d
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3213016
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>