hterm: add a line height padding option
This allows the user to force a little extra spacing between lines if
the default behavior is too tight for their preference.  In order to
support things like ASCII box drawings, we need glyphs to be close,
but users might be OK with the trade off of a few gaps.

Bug: https://crbug.com/1256106
Change-Id: I8b8eee81204c4abf0290513a0b2f8e2ede36d71d
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3216444
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>