hterm: make setProfile(id) a nop if id does not change
When terminal-ssh loads, it sets a background-color watcher on prefs to
set background of the top-most document which is used for styling tabs
in the SWA. The watcher does not continue after an ssh connection since
prefs is being recreated when nassh calls terminal.setProfile().
Since the 'default' profile is still being used, there is no need to
deactive and recreate prefs.

Bug: 1241262
Change-Id: Iabcfddc8784506a45e285ed7d09442edbc8810a7
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3229054
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>