terminal: adjust value matching test.
This is a followup to: https://crrev.com/c/apps/libapps/+/3233736

When the pref is set to a string (and not an int), this.value remains a
string.  Assert that current-value is set to the same value and not an
empty string.  Otherwise this test passes even without the fix to
findSelectedIndex_().

Change-Id: I203a5051c54195583a58ddda4063ba1b96a45d56
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3234998
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>