wasi-js-bindings: refine use of ArrayBufferView
The read/write funcs don't handle DataView, so reduce ArrayBufferView
down to TypedArray.  For the read funcs, just say we'll always return
Uint8Array (since we do) so callers don't have to get fancy.  Also
refine the read returns to reflect the code -- we allow buf to not be
defined, and we allow nread to be omitted if buf is provided.

Change-Id: I030e2548488f0f194b60992e172121f66a898fbe
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3498289
Reviewed-by: Emil Mikulic <easy@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>