nassh: mkdeps: setup wassh paths for development
This makes it easier to get code up & running inside unpacked nassh.
We don't dist anything yet though.

Also hack the wassh paths slightly so they work when loaded in situ
or when loaded via the extension.  This leads to a minor error if
you're looking at the console about a file not found, but it works.

Bug: https://crbug.com/1300798
Change-Id: If5027562e56e34d46e6a77470de1f239b7d20c67
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3509231
Reviewed-by: Emil Mikulic <easy@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>