nassh: corpv4: include ssh username in /connect call
To help the relay server with logging, include the ssh username.
This code is a bit lazy in that we pass down --proxy-user which is
the "relay username" even though dstUsername is supposed to be the
"ssh username", but we don't use --proxy-user otherwise here, and
it already defaults to the ssh username.

Bug: http://b/219762667
Change-Id: I1d5a977577afc5036fb36cb895ad123989ea8ddf
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3526473
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>