wassh: implement basic signal delivery
Drop the handlers API as it's incompatible with the current signal stubs
that WASI is providing (since it has its own internal handlers).

Add a new __wassh_signal_deliver callback for the JS side to call when
it has a signal to deliver.  This will look up the current handler and
then call it directly.  We don't support sigaction+SA_SIGINFO atm, but
OpenSSH doesn't use that, so isn't a big deal.  We can fake this though
easily in the future if a need arises.

On the JS side of things, plumbing signal logic is a little complicated
because of the worker interaction.  We can't send a message as it won't
be delivered (since the worker is always synchronous).  So we include it
when sending a syscall response back, and it takes care of calling into
the WASM code to process it.

At a high level, we're implementing signal delivery a bit like pthread
cancellation: it isn't truly synchronous, and doesn't truly interrupt
syscalls, but even POSIX doesn't guarantee that (signals get queued,
and then delivered when possible).  For now we only check this in the
poll_oneoff syscall since that's the main loop construct that OpenSSH
uses (it's constantly polling for fd's for events).  We could add it
to more syscalls if desired in the central syscall hook point, but
this works well enough -- terminal resizes are delivered immediately.

An alternate design that would avoid a lot of this complicated logic
would be based on signalfd, but that would require invasive changes
to OpenSSH, so let's punt on that idea for now.

One thing worth noting is that WASI defines a set of signal numbers,
but doesn't currently use them anywhere, and their numbers seem to be
different from musl's signal numbers.  The delivery mechanism here is
using musl's signal numbers.  If WASI ever figures out what it wants
to do with signals, it should be easy enough to adapt.

Bug: https://crbug.com/1307963
Change-Id: Ib8f3e01068556cbef69f4483dadd6ce7443a2af4
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3538105
Reviewed-by: Emil Mikulic <easy@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>