wassh: connect ssh-agent streams through
Plumb the ssh-agent logic through UNIX sockets similar to how standard
OpenSSH works, for the most part.  The NaCl code has been patching the
agent connect to using a fake TCP/IP address instead of UNIX sockets.
To that end, we detect the fake IP when connecting on the JS side, and
switch it to a UNIX socket.  Ugly, but gets the job done, and we'll be
dropping the custom OpenSSH patch once we drop the NaCl port.

Bug: https://crbug.com/1303492
Change-Id: Ib945d6b9df16b8d15dcf6c42edb3dc9fa2a072de
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3553464
Reviewed-by: Emil Mikulic <easy@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>