wassh: connect relay streams through
Plumb the existing nassh relay server logic through when using wassh.
It's difficult to determine, from the socket connect syscall, whether
the request is for the ssh system via the relay, or to a different
system (e.g. port forwarding).  So for now, we assume that the first
TCP/IP connect call is for the relay, and all others are for local
systems.  This should hopefully work well enough for now.

We can't do something simple like string compare the server settings
with the address in the connect call because the user's ssh_config
might have rewritten it slightly.

If this ends up not working well, we can revisit with a simple hack:
patch OpenSSH to pass unique protocol settings in ssh_connect_direct
when creating the socket.

Longer term, we should figure out how to hijack ProxyCommand to do
the relay setup since that's how gLinux does it, and how everyone
else in the world connects via relays.  But for now, we focus on the
logic working the same way as the NaCl plugin -- it forces the first
connect to the JS relay code, and then drops the rest to direct.

Bug: https://crbug.com/1310571
Change-Id: I5d8cd66960452e0bff693cc05e8e8b8d26a836cb
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3553465
Reviewed-by: Emil Mikulic <easy@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>