nassh: switch pages without adding to history
When redirecting via the relay, or selecting a profile, we've been
adding new entries to the current tab's history.  This inadvertently
allows for OS-level back/forward web navigation which can cause the
connection to immediately be lost.  Use Location's .replace() API so
we never generate new history entries.  This automatically lets the
browser disable forward/back navigation.

This doesn't work completely for in-place SSH URIs as we have no way
of clearing existing history, but this should be a good improvement
for normal extension users.

Bug: https://crbug.com/952121
Change-Id: I10d0bc7a06c90ffbd1aa06fda80f16e90f30e4a7
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3561279
Reviewed-by: Jason Lin <lxj@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>