terminal: update home page with a11y autofocus
Autofocus on first element "Add SSH" button so that chromevox will
always read something when the page opens.

Update messaging for when crostini.enabled is false.  This does not
necessarily mean that it has been disabled by admin.  Changed message to
state that Linux is not enabled and to see Developer Settings page where
more info can be shown.

Bug: b/223076712
Change-Id: Ifdb6caff2fd78dc5e49ce7cf4313054bd9a8a58e
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3574355
Reviewed-by: Jason Lin <lxj@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>