ssh_client: openssh: fix memory corruption with select()
OpenSSH was making assumptions about the internal C library fd_set layout
leading to memory corruption and crashes at runtime.  It happened to work
with NaCl, but not with WASI libc.

This isn't needed in newer versions of OpenSSH 8.9+ because they rewrote
this logic entirely, but we still need to patch older versions we use.

Bug: https://crbug.com/1312256
Change-Id: I27a1592233a4d06d711749d9697d730204b74a95
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3580620
Reviewed-by: Emil Mikulic <easy@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>