nassh: prefs: keep large border sizes from breaking settings page
If the screen border size is set to a very large value (e.g. 999999),
it causes the terminal preview to grow so large that the settings can't
be changed, including resetting the screen border size.  Add a limit to
the area to prevent any such change from breaking the preferences page.

Bug: https://crbug.com/1179808
Change-Id: Ic61527bc223a9385e21e3feeb5fedee4c483c2c7
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3586232
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>