http-server: fix --cors usage
The http-server has poor parsing logic and will consume any non-option
following it as a header value.  This was causing it to use the path as
a header option with the CORS HTTP headers.  Move it to the start of the
command line before the --address option so that doesn't happen.

This didn't show up for the most part because usually http-server is run
with the cwd already set to the directory we wanted to use.

Change-Id: Ie79742ab104ab62aff29fbb842ef27ebb53523dd
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3586234
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>