nassh: allow nassh_preferences_editor.html to run in System App
It can be helpful to run
chrome-untrusted://terminal/html/nassh_preferences_editor.html which
gives access to some things such as custom CSS which is not available in
the Terminal settings page.

chrome.runtime.sendMessage is not available in that env, so backup link
will not work, but this is not a big deal.

Change-Id: I0ad6f8d81170cc79260f22890b09dc65898f4316
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/3632045
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>